feat: 전체 크기 재조정
