+++ wingle/src/components/layout/Footer/index.tsx
@@ -1,3 +1,18 @@
+import styled from "styled-components";
+
+const Style = {
+  Wrapper: styled.div`
+    width: 500px;
+    height: 270px;
+    background-color: #fcfcfc;
+    border-top: 1px solid #eeeef2;
+  `,
+};
+
 export default function Footer() {
-  return <>푸터</>;
+  return (
+    <>
+      <Style.Wrapper>얍얍</Style.Wrapper>
+    </>
+  );
 }
