+++ wingle/src/pages/_app.tsx
@@ -4,16 +4,28 @@ import { ThemeProvider } from "styled-components";
 import { GlobalStyle } from "../styles/global-style";
 import { theme } from "../styles/theme";
 
+import styled from "styled-components";
+const Style = {
+  Wrapper: styled.div`
+    width: 500px;
+    height: 951px;
+    background-color: white;
+  `,
+};
+
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <>
       <Head>
         <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <title>boilerplate</title>
+        <title>국제교류 온라인 플랫폼, Wingle</title>
+        <link rel="icon" href="/logo_favicon.jpeg" />
       </Head>
       <GlobalStyle />
       <ThemeProvider theme={theme}>
-        <Component {...pageProps} />
+        <Style.Wrapper>
+          <Component {...pageProps} />
+        </Style.Wrapper>
       </ThemeProvider>
     </>
   );
