+++ wingle/src/pages/index.tsx
@@ -1,24 +1,13 @@
 import Head from "next/head";
-import styled from "styled-components";
-
-const Style = {
-  Wrapper: styled.div`
-    width: 500px;
-    height: 951px;
-    background-color: white;
-  `,
-};
 
 export default function Home() {
   return (
     <>
       <Head>
-        <title>국제교류 온라인 플랫폼, Wingle</title>
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/logo_favicon.jpeg" />
       </Head>
-      <Style.Wrapper>내용</Style.Wrapper>
+      <main></main>
     </>
   );
 }
